requirements.txt
python == 3.12
google-genai == 1.21.1 # check https://pypi.org/project/google-genai/ for SDK installation
python-dotenv == 1.1.1
